Make user handling unicode tolerant
Previously it was unclear whether or not support for unicode was
present here. This was mainly because either our wrappers or the
python-ldap module would error when they saw a 'unicode' object.

This commit ensures that the user class can be passed a unicode
instance, but internally uses and returns 'str's. This ensures
that the data can be passed into the database properly while also
enabling clients of this library to pass unicode strings in.

All property sets on user instances should now support being given
unicode. The new_username method now also supports being given either
type, where previously it really wanted to be given unicode.

This change should be transparent to 'str' users while simplifying
things for 'unicode' users (nemesis being the main one).

Change-Id: I4e8211ee1f5c4cdfae496669ef9e33c074a09baf